fix(core): more gracefully handle dom and errors in plugin compiler
Fixes #306
